chore: simplify the advanced search
