+++ frontend/src/views/InstanceDetail.vue
@@ -182,10 +182,11 @@ const state = reactive<LocalState>({
 
 const route = useRoute();
 
-const scopeOptions = useCommonSearchScopeOptions(
-  computed(() => state.params),
-  [...CommonFilterScopeIdList, "project", "database-label"]
-);
+const scopeOptions = useCommonSearchScopeOptions([
+  ...CommonFilterScopeIdList,
+  "project",
+  "database-label",
+]);
 
 watch(
   () => route.hash,
